+++ src/Logary.Tests/Literate.fs
@@ -31,8 +31,8 @@ let tokenisation =
       let expected =
         [ "-", MessageTemplates.Formatting.Literate.Subtext // newline
           "System.Exception: Top level exn", MessageTemplates.Formatting.Literate.Text
-          "Logary.Tests.LiterateConsole.throwAnotherExn@65(Exception inner)", MessageTemplates.Formatting.Literate.Subtext
-          "--- End of inner exception stack trace ---", MessageTemplates.Formatting.Literate.Punctuation
+          "Logary.Tests.LiterateConsole.throwAnotherExn@19(Exception inner)", MessageTemplates.Formatting.Literate.Subtext
+          "--- End of exception stack trace ---", MessageTemplates.Formatting.Literate.Punctuation
           "System.Exception: Bad things going on", MessageTemplates.Formatting.Literate.Text
           "Logary.Tests.Utils.innermost[a]()", MessageTemplates.Formatting.Literate.Subtext
         ]
@@ -46,14 +46,17 @@ let tokenisation =
       let expected =
         [ "-", MessageTemplates.Formatting.Literate.Subtext // newline
           "System.AggregateException: Outer aggregate exception (Bad things going on) (Bad things going on)", MessageTemplates.Formatting.Literate.Text
-          "Logary.Tests.LiterateConsole.throwAggrExn@69(Exception inner)", MessageTemplates.Formatting.Literate.Subtext
-          "--- End of inner exception stack trace ---", MessageTemplates.Formatting.Literate.Punctuation
+          "Logary.Tests.LiterateConsole.throwAggrExn@23(Exception inner1, Exception inner2)", MessageTemplates.Formatting.Literate.Subtext
+          "--- End of exception stack trace ---", MessageTemplates.Formatting.Literate.Punctuation
+          "System.Exception: Bad things going on", MessageTemplates.Formatting.Literate.Text
+          "Logary.Tests.Utils.innermost[a]()", MessageTemplates.Formatting.Literate.Subtext
+          "--- End of exception stack trace ---", MessageTemplates.Formatting.Literate.Punctuation
           "System.Exception: Bad things going on", MessageTemplates.Formatting.Literate.Text
           "Logary.Tests.Utils.innermost[a]()", MessageTemplates.Formatting.Literate.Subtext
         ]
       Message.event Warn "My error" |> Message.addExn exnOuter
         |> Literate.tokeniseExceptions Culture.invariant "-"
-        |> Expect.sequenceEqual "Has correct output" expected
+        |> Expect.sequenceContainsOrder "Has correct output" expected
   ]
 
 module LiterateTesting =