+++ RELEASE_NOTES.md
@@ -1,6 +1,7 @@
 #### 5.0.0-beta.21
 * Fix #357 — for simple exns
 * Fix #357 — for AggregateException
+* Rename `StacktraceParse.InnerDelim` -> `StacktraceParse.StacktraceDelim`, since it's being used for non-inner stacktrace delimitations.
 
 #### 5.0.0-beta.20
 * Enable access to Suave.HttpContext for HTTP ingestion